// apps/web/src/types/opencode.gen.ts
export type ProviderAuthError = {
  name: 'ProviderAuthError';
  data: {
    providerID: string;
    message: string;
  };
};

export type UnknownError = {
  name: 'UnknownError';
  data: {
    message: string;
  };
};

export type MessageOutputLengthError = {
  name: 'MessageOutputLengthError';
  data: {
    [key: string]: unknown;
  };
};

export type MessageAbortedError = {
  name: 'MessageAbortedError';
  data: {
    message: string;
  };
};

export type StructuredOutputError = {
  name: 'StructuredOutputError';
  data: {
    message: string;
    retries: number;
  };
};

export type ContextOverflowError = {
  name: 'ContextOverflowError';
  data: {
    message: string;
    responseBody?: string;
  };
};

export type ApiError = {
  name: 'APIError';
  data: {
    message: string;
    statusCode?: number;
    isRetryable: boolean;
    responseHeaders?: {
      [key: string]: string;
    };
    responseBody?: string;
    metadata?: {
      [key: string]: string;
    };
  };
};

export type AssistantMessage = {
  id: string;
  sessionID: string;
  role: 'assistant';
  time: {
    created: number;
    completed?: number;
  };
  error?:
    | ProviderAuthError
    | UnknownError
    | MessageOutputLengthError
    | MessageAbortedError
    | StructuredOutputError
    | ContextOverflowError
    | ApiError;
  parentID: string;
  modelID: string;
  providerID: string;
  mode: string;
  agent: string;
  path: {
    cwd: string;
    root: string;
  };
  summary?: boolean;
  cost: number;
  tokens: {
    total?: number;
    input: number;
    output: number;
    reasoning: number;
    cache: {
      read: number;
      write: number;
    };
  };
  structured?: unknown;
  variant?: string;
  finish?: string;
};